find my fund 1/CODE/requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
scikit-learn
pandas
numpy
gradio
tqdm
fastapi
uvicorn
peft
accelerate 